Changed GameMovableDrive to an abstract class instead of an interface.
It allows to set the MoveStrategy and the MoveBlockerChecker when using
the type GameMovableDriver.
